Split option_map and declaration in separate files
