+++ libvast/CMakeLists.txt
@@ -55,6 +55,7 @@ set(libvast_sources
   src/key.cpp
   src/null_bitmap.cpp
   src/operator.cpp
+  src/option_declaration_set.cpp
   src/option_map.cpp
   src/pattern.cpp
   src/port.cpp
@@ -227,6 +228,7 @@ set(tests
   test/main.cpp
   test/mmapbuf.cpp
   test/offset.cpp
+  test/option_declaration_set.cpp
   test/option_map.cpp
   test/parseable.cpp
   test/pattern.cpp
