+++ libvast/test/option_map.cpp
@@ -17,8 +17,7 @@
 #include <limits>
 
 #include "vast/option_map.hpp"
-
-#include "vast/detail/string.hpp"
+#include "vast/option_declaration_set.hpp"
 
 using namespace vast;
 
@@ -32,64 +31,6 @@ struct fixture {
 } // namespace <anonymous>
 
 FIXTURE_SCOPE(command_tests, fixture)
-
-TEST(declaration adding) {
-  CHECK(decl.add("flag,fabc", "this is a flag", false));  
-  CHECK(decl.add("str,s", "this is a string", std::string("")));  
-  CHECK(decl.add("test-int", "this is an int", 1));  
-  CHECK(!decl.add(",x", "using only a short name", 1));  
-  CHECK(!decl.add("flag", "using the same long name again", false));  
-  CHECK_EQUAL(decl.size(), 3u + 1u); //3 options added + the help option
-  auto x = decl.find("help");
-  MESSAGE("Test help option");
-  CHECK_EQUAL(x->long_name(), "help");
-  REQUIRE_EQUAL(x->short_names().size(), 2u);
-  CHECK_EQUAL(x->short_names()[0], 'h');
-  CHECK_EQUAL(x->short_names()[1], '?');
-  CHECK_EQUAL(x->description(), "print this text");
-  CHECK_EQUAL(x->has_argument(), false);
-  REQUIRE(x);
-  MESSAGE("Test flag option");
-  x = decl.find("flag");
-  REQUIRE(x);
-  CHECK_EQUAL(x->long_name(), "flag");
-  REQUIRE_EQUAL(x->short_names().size(), 4u);
-  CHECK_EQUAL(x->short_names()[0], 'f');
-  CHECK_EQUAL(x->short_names()[3], 'c');
-  CHECK_EQUAL(x->description(), "this is a flag");
-  CHECK_EQUAL(x->has_argument(), false);
-  MESSAGE("Test string option");
-  x = decl.find("str");
-  REQUIRE(x); 
-  CHECK_EQUAL(x->long_name(), "str");
-  CHECK_EQUAL(x->has_argument(), true);
-}
-
-TEST(data type parsing) {
-  using option = option_declaration_set::option_declaration;
-  auto check_option
-    = [](const option& opt, const std::string& str, auto expected_value) {
-        auto d = opt.parse(str);
-        REQUIRE(d);
-        CHECK_EQUAL(*d, expected_value);
-      };
-  auto check_fail_option = [](const option& opt, const std::string& str) {
-    auto d = opt.parse(str);
-    CHECK(!d);
-  }; 
-  CHECK(decl.add("int", "", 1));  
-  CHECK(decl.add("string", "", ""));  
-  MESSAGE("Test int");
-  auto x = decl.find("int");
-  check_option(*x, "2", 2);
-  check_option(*x, "0", 0);
-  check_option(*x, "-2", -2);
-  check_fail_option(*x, "X");
-  MESSAGE("Test string");
-  x = decl.find("string");
-  check_option(*x, "2", "2");
-  check_option(*x, "this is a test", "this is a test");
-}
 
 TEST(option map handling) {
   auto num = 42;
@@ -113,7 +54,7 @@ TEST(option map handling) {
   CHECK_EQUAL(y, num);
 }
 
-TEST(parse cli) {
+TEST(cli parsing) {
   auto split = [](const std::string& str) {
     std::vector<std::string> result;
     auto xs = detail::split(str.begin(), str.end(), " ");
